// ops-deck-0013/src/types/db.ts
export type TestStatus =
  | 'not_tested' | 'planned' | 'preparing' | 'ready' | 'testing' | 'winner' | 'stopped'